requirements.txt
fastapi
uvicorn[standard]
httpx
mcp[cli]
PyGithub
PyPDF2
requests 
python-dotenv
websockets
langchain
langgraph
pydantic
python-multipart